fix(payments): isolate anonymous claim storage
